Centralize profile picture utilities
