+++ app/utils/refreshCachedUserPhoto.js
@@ -1,4 +1,4 @@
-import { getUserPfp } from "./conversationsFunctions";
+import { getUserPfp } from "./uploadFile";
 
 export async function refreshCachedUserPhoto(uid, updateCachedUserLogin) {
   const photo = await getUserPfp(uid);
